# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel audit",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a list of documents",
                "operationId": "8eb2403bd51118a9afdbec01a985b9ba",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a specific document",
                "operationId": "eaa00bafa49c865f4222dbbf894cce22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "documentName": {
                                            "type": "string"
                                        },
                                        "fields": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "field_seq": {
                                                        "type": "integer"
                                                    },
                                                    "is_mandatory": {
                                                        "type": "integer"
                                                    },
                                                    "field_type": {
                                                        "type": "integer"
                                                    },
                                                    "field_name": {
                                                        "type": "string"
                                                    },
                                                    "document_id": {
                                                        "type": "integer"
                                                    },
                                                    "select_values": {
                                                        "type": "string",
                                                        "format": "json"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/documents/create": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "new document",
                "operationId": "9f65012a800ae50659e412e1eebce997",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "document_name"
                                ],
                                "properties": {
                                    "document_name": {
                                        "type": "string",
                                        "example": "Apple"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "this id good fruit"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "3000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful create document"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        }
    },
    "components": {
        "headers": {
            "Accept": {
                "description": "Accept header",
                "schema": {
                    "type": "string",
                    "default": "application/json",
                    "enum": [
                        "application/json"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Documents",
            "description": "Documents"
        }
    ]
}